Refactor MPEG parsing logic into `MpegAudio` class to avoid duplicating code.
